+++ RELEASE-NOTES.md
@@ -23,3 +23,4 @@
 - add wallet address to wallet button
 - react rendering performance upgrades and UI improvements
 - block number and timestamp UI updates
+- added `SLink` component for internal and external links